modularize: split src/ files over 400 lines into focused modules
Four src/ files crossed the 400-line threshold after the #39 and
profiling/timing work landed. This commit splits each into focused
sibling modules that can be read and tested in isolation, while
preserving every existing public import path via thin barrel files
so no test file or external caller has to change.

Before                                        After                                            Lines
────────────────────────────────────────────  ──────────────────────────────────────────────  ─────
src/tools/tool-helpers.ts              (702)  tool-helpers.ts (barrel re-export)                 63
                                              tool-provenance.ts (null-proto registry)           70
                                              tool-envelope.ts (types + schemas + ctors)        146
                                              tool-errors.ts (ToolInvocationError + helpers)     85
                                              tool-gates.ts (session/goal gates + grouping)    165
                                              tool-registration.ts (register*Tool + wrappers) 309

src/session/load-tool-registration.ts  (548)  load-tool-registration.ts (orchestrator)           37
                                              load-tool-shared.ts (error helpers + resolver)   136
                                              register-agda-load.ts                            217
                                              register-agda-load-no-metas.ts                   127
                                              register-agda-typecheck.ts                       148

src/agda/session.ts                    (534)  session.ts (class + lifecycle)                   369
                                              binary-discovery.ts (findAgdaBinary)              21
                                              goal-merging.ts (pure mergeGoals)                 41
                                              session-constants.ts (NOT_FOUND_RESULT)           28
                                              session-load-impl.ts (runLoad, runLoadNoMetas)   200

src/tools/reporting-tools.ts           (438)  reporting-tools.ts (orchestrator)                 33
                                              reporting-schemas.ts (zod + render helpers)      118
                                              register-tools-catalog.ts                         48
                                              register-protocol-parity.ts                       82
                                              register-bug-bundles.ts                          210

The tool-helpers split preserves the full public import surface via
a barrel re-export in tool-helpers.ts — every `import … from
"./tool-helpers"` in src/ and test/ keeps working without change.
New code should import directly from the focused modules where
possible.

The session.ts refactor makes three load-history fields
(lastLoadedMtime, lastClassification, lastLoadedAt) package-internal
(removed the `private` modifier) so sibling session-load-impl.ts can
update them from runLoad / runLoadNoMetas. External consumers still
read via the getters (isFileStale, getLastClassification,
getLastLoadedAt), which are unchanged. AgdaSession also gains a
public iotcmFor(filePath, agdaCmd) helper — an explicit-file
variant of the pre-existing private buildIotcm — that session-load-
impl.ts uses to construct Cmd_load without first mutating
currentFile (which would race with ensureProcess()'s stale-state
reset).

The load-tool-registration split extracts a small
validateProfileOptionsOrError helper to load-tool-shared.ts so each
of the three register-agda-*.ts files can handle the profile-option
validation fast-fail in a single line rather than duplicating the
error envelope shape.

After this commit no file in src/ exceeds 400 lines. The biggest is
session.ts at 369; the next biggest is response-schemas.ts at 310.

Verification:
- npx tsc -p tsconfig.json --noEmit: clean on first try.
- npx vitest run test/unit: 370 passed (66 files).
- RUN_AGDA_INTEGRATION=1 npx vitest run test/integration: 144 passed,
  2 skipped.
- npm run build: clean.

No behaviour change — every extracted function has the same
signature and semantics as the original inline code.

Refs #39.
