+++ src/tools/tool-helpers.ts
@@ -1,702 +1,63 @@
 // MIT License — see LICENSE
 //
-// Shared helpers for MCP tool handlers: staleness warnings, goal
-// validation, manifest registration, and structured output envelopes.
-
-import { McpServer } from "@modelcontextprotocol/sdk/server/mcp.js";
-import type { ToolAnnotations } from "@modelcontextprotocol/sdk/types.js";
-import { z } from "zod";
-
-import type { AgdaSession } from "../agda-process.js";
-import { PathSandboxError } from "../repo-root.js";
-import type { ToolCategory } from "./manifest.js";
-import { registerManifestEntry } from "./manifest.js";
-
-export interface ToolDiagnostic {
-  severity: "error" | "warning" | "info";
-  message: string;
-  code?: string;
-}
-
-export interface ToolEnvelope<T extends Record<string, unknown>> {
-  tool: string;
-  ok: boolean;
-  classification: string;
-  summary: string;
-  data: T;
-  diagnostics: ToolDiagnostic[];
-  stale?: boolean;
-  provenance?: Record<string, unknown>;
-  /** Wall-clock time for the tool invocation in milliseconds. */
-  elapsedMs?: number;
-}
-
-// ── Global provenance ─────────────────────────────────────────────────
+// Barrel re-export for tool helpers.
 //
-// Keys registered here are merged into every tool envelope's `provenance`
-// field by okEnvelope / errorEnvelope. Used for process-wide metadata
-// (Agda version, server version, toolchain hash) that every response
-// should carry so an agent doesn't have to re-ask. Tool-specific
-// provenance keys from the call site always override global keys with
-// the same name.
+// This module used to hold all of: ToolEnvelope/ToolResult types,
+// diagnostic constructors, provenance registry, error translation,
+// session/goal gates, and register*Tool wrappers. It grew past 700
+// lines so it was split into five focused modules; this file stays
+// as a thin re-export so every existing `import … from "./tool-helpers"`
+// call site in src/ and test/ keeps working unchanged.
 //
-// SECURITY: the registry and the merge result are both null-prototype
-// objects. This prevents a future caller from writing to `__proto__` /
-// `constructor` / other special property slots via registerGlobalProvenance
-// and affecting lookups on unrelated objects (CWE-1321, "Prototype
-// Pollution"). It also means JSON.stringify is the only observer of these
-// objects, so there's no accidental inheritance from Object.prototype.
+// Internal modules (import directly from these in new code):
+//   - tool-provenance.ts   — global provenance registry
+//   - tool-envelope.ts     — envelope types, schemas, constructors
+//   - tool-errors.ts       — ToolInvocationError, missingPathToolError
+//   - tool-gates.ts        — session + goal gates, staleness warning
+//   - tool-registration.ts — register*Tool wrappers, wrap*Handler
 
-const UNSAFE_KEYS = new Set(["__proto__", "prototype", "constructor"]);
+export {
+  registerGlobalProvenance,
+  clearGlobalProvenance,
+  mergeProvenance,
+  type ProvenanceRecord,
+} from "./tool-provenance.js";
 
-type ProvenanceRecord = Record<string, unknown>;
+export {
+  diagnosticSchema,
+  errorDiagnostic,
+  errorEnvelope,
+  infoDiagnostic,
+  makeToolResult,
+  okEnvelope,
+  text,
+  toolEnvelopeSchema,
+  warningDiagnostic,
+  type ToolDiagnostic,
+  type ToolEnvelope,
+  type ToolResult,
+} from "./tool-envelope.js";
 
-function createProvenanceRecord(): ProvenanceRecord {
-  return Object.create(null) as ProvenanceRecord;
-}
+export {
+  ToolInvocationError,
+  makeTextToolErrorResult,
+  missingPathToolError,
+  toToolInvocationError,
+} from "./tool-errors.js";
 
-const globalProvenance: ProvenanceRecord = createProvenanceRecord();
+export {
+  groupDiagnosticsByFile,
+  sessionErrorStateGate,
+  stalenessWarning,
+  validateGoalId,
+} from "./tool-gates.js";
 
-export function registerGlobalProvenance(key: string, value: unknown): void {
-  if (typeof key !== "string" || key.length === 0 || UNSAFE_KEYS.has(key)) {
-    return;
-  }
-  if (value === undefined || value === null) {
-    delete globalProvenance[key];
-    return;
-  }
-  globalProvenance[key] = value;
-}
-
-export function clearGlobalProvenance(): void {
-  for (const key of Object.keys(globalProvenance)) {
-    delete globalProvenance[key];
-  }
-}
-
-function copyOwnStringKeys(
-  dst: ProvenanceRecord,
-  src: ProvenanceRecord | undefined,
-): void {
-  if (!src) return;
-  for (const key of Object.keys(src)) {
-    if (UNSAFE_KEYS.has(key)) continue;
-    if (!Object.prototype.hasOwnProperty.call(src, key)) continue;
-    dst[key] = src[key];
-  }
-}
-
-function mergeProvenance(
-  local: ProvenanceRecord | undefined,
-): ProvenanceRecord | undefined {
-  const globalKeys = Object.keys(globalProvenance);
-  const hasLocal = local !== undefined && Object.keys(local).length > 0;
-  if (globalKeys.length === 0 && !hasLocal) {
-    return undefined;
-  }
-  const merged = createProvenanceRecord();
-  copyOwnStringKeys(merged, globalProvenance);
-  copyOwnStringKeys(merged, local);
-  return merged;
-}
-
-export class ToolInvocationError<T extends Record<string, unknown> = Record<string, unknown>> extends Error {
-  classification: string;
-  diagnostics: ToolDiagnostic[];
-  data: T;
-  text?: string;
-
-  constructor(args: {
-    message: string;
-    classification?: string;
-    diagnostics?: ToolDiagnostic[];
-    data?: T;
-    text?: string;
-  }) {
-    super(args.message);
-    this.name = "ToolInvocationError";
-    this.classification = args.classification ?? "tool-error";
-    this.diagnostics = args.diagnostics ?? [errorDiagnostic(args.message)];
-    this.data = args.data ?? ({} as T);
-    this.text = args.text;
-  }
-}
-
-type ToolResult<T extends Record<string, unknown> = Record<string, unknown>> = {
-  content: { type: "text"; text: string }[];
-  structuredContent: ToolEnvelope<T>;
-  isError?: boolean;
-};
-
-export const diagnosticSchema = z.object({
-  severity: z.enum(["error", "warning", "info"]),
-  message: z.string(),
-  code: z.string().optional(),
-});
-
-const envelopeBaseSchema = z.object({
-  tool: z.string(),
-  ok: z.boolean(),
-  classification: z.string(),
-  summary: z.string(),
-  diagnostics: z.array(diagnosticSchema),
-  stale: z.boolean().optional(),
-  provenance: z.record(z.string(), z.unknown()).optional(),
-  elapsedMs: z.int().nonnegative().optional(),
-});
-
-export function toolEnvelopeSchema(
-  dataSchema: z.ZodTypeAny,
-): z.ZodTypeAny {
-  return envelopeBaseSchema.extend({
-    data: dataSchema,
-  });
-}
-
-export function makeToolResult<T extends Record<string, unknown>>(
-  envelope: ToolEnvelope<T>,
-  text?: string,
-): ToolResult<T> {
-  return {
-    content: [{ type: "text" as const, text: text ?? envelope.summary }],
-    structuredContent: envelope,
-    isError: !envelope.ok,
-  };
-}
-
-export function infoDiagnostic(message: string, code?: string): ToolDiagnostic {
-  return { severity: "info", message, code };
-}
-
-export function warningDiagnostic(message: string, code?: string): ToolDiagnostic {
-  return { severity: "warning", message, code };
-}
-
-export function errorDiagnostic(message: string, code?: string): ToolDiagnostic {
-  return { severity: "error", message, code };
-}
-
-export function okEnvelope<T extends Record<string, unknown>>(args: {
-  tool: string;
-  summary: string;
-  data: T;
-  classification?: string;
-  diagnostics?: ToolDiagnostic[];
-  stale?: boolean;
-  provenance?: Record<string, unknown>;
-  elapsedMs?: number;
-}): ToolEnvelope<T> {
-  return {
-    tool: args.tool,
-    ok: true,
-    classification: args.classification ?? "ok",
-    summary: args.summary,
-    data: args.data,
-    diagnostics: args.diagnostics ?? [],
-    stale: args.stale,
-    provenance: mergeProvenance(args.provenance),
-    elapsedMs: args.elapsedMs,
-  };
-}
-
-export function errorEnvelope<T extends Record<string, unknown>>(args: {
-  tool: string;
-  summary: string;
-  data: T;
-  classification?: string;
-  diagnostics?: ToolDiagnostic[];
-  stale?: boolean;
-  provenance?: Record<string, unknown>;
-  elapsedMs?: number;
-}): ToolEnvelope<T> {
-  return {
-    tool: args.tool,
-    ok: false,
-    classification: args.classification ?? "tool-error",
-    summary: args.summary,
-    data: args.data,
-    diagnostics: args.diagnostics ?? [errorDiagnostic(args.summary)],
-    stale: args.stale,
-    provenance: mergeProvenance(args.provenance),
-    elapsedMs: args.elapsedMs,
-  };
-}
-
-export function missingPathToolError(kind: "file" | "directory", path: string): ToolInvocationError<{ path: string }> {
-  const message = `${kind === "file" ? "File" : "Directory"} not found: ${path}`;
-  return new ToolInvocationError({
-    message,
-    classification: "not-found",
-    diagnostics: [errorDiagnostic(message, "not-found")],
-    data: { path },
-  });
-}
-
-/** Return a staleness warning if the loaded file was modified on disk. */
-export function stalenessWarning(session: AgdaSession): string {
-  if (session.isFileStale()) {
-    return "**Warning:** File modified since last load — results may be stale. Run `agda_load` to refresh.\n\n";
-  }
-  return "";
-}
-
-// Matches the leading file path in an Agda diagnostic. Same family of
-// patterns as parse-load-responses.ts#ERROR_LOCATION_PATTERN, but we
-// capture the filename (group 1) instead of the line number because
-// the caller cares about which file the diagnostic is *about*, not
-// where in it the error was. Deliberately loose on the path so that
-// both absolute and relative forms match, and both .agda and .lagda
-// variants are recognized.
-const DIAGNOSTIC_FILE_PATTERN = /([^\s:]+?\.(?:lagda\.md|lagda|agda)):\d+/u;
-
-/**
- * Group diagnostic strings by the file path mentioned at their head.
- *
- * §1.2 from docs/bug-reports/agent-ux-observations.md: an agent that
- * calls agda_metas after a seemingly-clean agda_load can see a
- * well-formatted error from a transitive dependency of the loaded
- * file. Today those messages are a flat string array with no source
- * attribution, so the agent can't distinguish "the loaded file is
- * broken" from "a dependency is broken". Grouping the messages by the
- * file they reference — with a null bucket for messages whose text
- * didn't carry a parseable file path — gives that attribution back.
- *
- * Ordering: the returned array preserves insertion order (the order
- * files were first seen in the input), which matches the order in
- * which Agda emitted the diagnostics.
- */
-export function groupDiagnosticsByFile(
-  messages: readonly string[],
-): Array<{ file: string | null; messages: string[] }> {
-  const groups = new Map<string | null, string[]>();
-  for (const message of messages) {
-    if (typeof message !== "string" || message.length === 0) continue;
-    const match = DIAGNOSTIC_FILE_PATTERN.exec(message);
-    const file = match ? match[1] : null;
-    const bucket = groups.get(file);
-    if (bucket) {
-      bucket.push(message);
-    } else {
-      groups.set(file, [message]);
-    }
-  }
-  return Array.from(groups, ([file, msgs]) => ({ file, messages: msgs }));
-}
-
-/**
- * Session-error gate for query-style tools.
- *
- * When the session's most recent load returned a "type-error"
- * classification, query tools like agda_why_in_scope / agda_infer /
- * agda_compute / agda_search_about / agda_show_module cannot produce
- * a meaningful answer — Agda's interactive process will echo the
- * previous load error as the "result" of the query, which gets
- * embedded in a happy-path payload and silently mislead the caller
- * (observations doc §1.3). Query-tool callbacks should call this gate
- * before invoking Agda: if it returns non-null, return the value
- * immediately; otherwise proceed.
- *
- * The gate is conservative: it only triggers on `lastClassification ===
- * "type-error"`. `ok-with-holes` loads, `ok-complete` loads, and the
- * no-load-yet state all let the query run as before.
- */
-export function sessionErrorStateGate<T extends Record<string, unknown>>(
-  session: AgdaSession,
-  tool: string,
-  emptyData: T,
-): ToolResult<T> | null {
-  const lastClassification = session.getLastClassification?.() ?? null;
-  if (lastClassification !== "type-error") {
-    return null;
-  }
-  const loadedFile = session.getLoadedFile?.() ?? null;
-  const fileHint = loadedFile ? ` (loaded file: ${loadedFile})` : "";
-  const summary =
-    `${tool} is unavailable: the session's most recent load failed `
-    + `with classification 'type-error'${fileHint}. `
-    + `Fix the load errors first, then re-run agda_load before issuing queries.`;
-  return makeToolResult(
-    errorEnvelope({
-      tool,
-      summary,
-      classification: "unavailable",
-      data: emptyData,
-      diagnostics: [errorDiagnostic(summary, "session-unavailable")],
-    }),
-    summary,
-  );
-}
-
-/** MCP text content helper. */
-export function text(t: string): ToolResult<{ text: string }> {
-  return makeToolResult(
-    okEnvelope({
-      tool: "text",
-      summary: t,
-      data: { text: t },
-    }),
-    t,
-  );
-}
-
-/**
- * Validate a goalId against the session's current goals.
- * Returns an error response if invalid, or null if valid.
- */
-export function validateGoalId(
-  session: AgdaSession,
-  goalId: number,
-  tool = "unknown",
-): ToolResult | null {
-  const loaded = session.getLoadedFile();
-  if (!loaded) {
-    return makeToolResult(
-      errorEnvelope({
-        tool,
-        summary: "No file loaded. Call `agda_load` first.",
-        classification: "no-loaded-file",
-        data: { goalId },
-      }),
-    );
-  }
-  const ids = session.getGoalIds();
-  if (!ids.includes(goalId)) {
-    const available = ids.length > 0
-      ? ids.map((id) => `?${id}`).join(", ")
-      : "(none)";
-    return makeToolResult(
-      errorEnvelope({
-        tool,
-        summary: `Invalid goal ID ?${goalId}. Available goals: ${available}`,
-        classification: "invalid-goal",
-        data: { text: "", goalId, availableGoalIds: ids },
-        diagnostics: [
-          errorDiagnostic(
-            `Invalid goal ID ?${goalId}. Available goals: ${available}`,
-            "invalid-goal",
-          ),
-          infoDiagnostic(
-            "Run `agda_load` to refresh goals after modifying the file.",
-            "reload-hint",
-          ),
-        ],
-      }),
-      `Invalid goal ID ?${goalId}. Available goals: ${available}\n\n` +
-      "Hint: Run `agda_load` to refresh goals after modifying the file.",
-    );
-  }
-  return null;
-}
-
-function toToolInvocationError(err: unknown): ToolInvocationError {
-  if (err instanceof ToolInvocationError) {
-    return err;
-  }
-
-  if (err instanceof PathSandboxError) {
-    return new ToolInvocationError({
-      message: err.message,
-      classification: "invalid-path",
-      diagnostics: [errorDiagnostic(err.message, "invalid-path")],
-      data: { path: err.targetPath },
-    });
-  }
-
-  const message = err instanceof Error ? err.message : String(err);
-  return new ToolInvocationError({ message: `Error: ${message}` });
-}
-
-function makeTextToolErrorResult(
-  tool: string,
-  err: unknown,
-  defaultData: Record<string, unknown>,
-): ToolResult<Record<string, unknown>> {
-  const toolError = toToolInvocationError(err);
-  return makeToolResult(
-    errorEnvelope({
-      tool,
-      summary: toolError.message,
-      classification: toolError.classification,
-      data: { ...defaultData, ...toolError.data },
-      diagnostics: toolError.diagnostics,
-    }),
-    toolError.text ?? toolError.message,
-  );
-}
-
-/**
- * Wrap a session tool handler with staleness warning and error handling.
- * The handler returns a complete structured envelope.
- * Automatically measures wall-clock time and sets elapsedMs on the envelope
- * if the handler did not already set it.
- */
-export function wrapStructuredHandler<T extends Record<string, unknown>>(
-  tool: string,
-  session: AgdaSession,
-  handler: () => Promise<{ envelope: ToolEnvelope<T>; text?: string }>,
-): () => Promise<ToolResult<Record<string, unknown>>> {
-  return async () => {
-    const startMs = performance.now();
-    try {
-      const result = await handler();
-      if (result.envelope.elapsedMs === undefined) {
-        result.envelope.elapsedMs = Math.round(performance.now() - startMs);
-      }
-      return makeToolResult(result.envelope, result.text);
-    } catch (err) {
-      return makeTextToolErrorResult(tool, err, {});
-    }
-  };
-}
-
-/**
- * Backward-compatible wrapper for existing text-only handlers.
- * New code should prefer wrapStructuredHandler.
- */
-export function wrapHandler(
-  session: AgdaSession,
-  handler: () => Promise<string>,
-): () => Promise<ToolResult<Record<string, unknown>>> {
-  return wrapStructuredHandler("text-tool", session, async () => {
-    const warn = stalenessWarning(session);
-    const body = await handler();
-    const textValue = warn + body;
-    return {
-      envelope: okEnvelope({
-        tool: "text-tool",
-        summary: body,
-        data: { text: body },
-        stale: session.isFileStale() || undefined,
-      }),
-      text: textValue,
-    };
-  });
-}
-
-/**
- * Wrap a goal-based tool handler with validation, staleness warning,
- * and error handling. The handler returns a complete structured envelope.
- * Automatically measures wall-clock time and sets elapsedMs on the envelope
- * if the handler did not already set it.
- */
-export function wrapStructuredGoalHandler<A extends Record<string, unknown>, T extends Record<string, unknown>>(
-  tool: string,
-  session: AgdaSession,
-  handler: (args: A & { goalId: number }) => Promise<{ envelope: ToolEnvelope<T>; text?: string }>,
-): (args: A & { goalId: number }) => Promise<ToolResult<Record<string, unknown>>> {
-  return async (args) => {
-    const startMs = performance.now();
-    const invalid = validateGoalId(session, args.goalId, tool);
-    if (invalid) return invalid;
-    try {
-      const result = await handler(args);
-      result.envelope.elapsedMs ??= Math.round(performance.now() - startMs);
-      return makeToolResult(result.envelope, result.text);
-    } catch (err) {
-      return makeTextToolErrorResult(tool, err, { text: "", goalId: args.goalId });
-    }
-  };
-}
-
-/**
- * Backward-compatible wrapper for existing text-only goal handlers.
- * New code should prefer wrapStructuredGoalHandler.
- */
-export function wrapGoalHandler<A extends Record<string, unknown>>(
-  session: AgdaSession,
-  handler: (args: A & { goalId: number }) => Promise<string>,
-): (args: A & { goalId: number }) => Promise<ToolResult<Record<string, unknown>>> {
-  return wrapStructuredGoalHandler("goal-tool", session, async (args) => {
-    const warn = stalenessWarning(session);
-    const body = await handler(args);
-    const textValue = warn + body;
-    return {
-      envelope: okEnvelope({
-        tool: "goal-tool",
-        summary: body,
-        data: { text: body, goalId: args.goalId },
-        stale: session.isFileStale() || undefined,
-      }),
-      text: textValue,
-    };
-  });
-}
-
-export function registerStructuredTool(args: {
-  server: McpServer;
-  name: string;
-  description: string;
-  category: ToolCategory;
-  protocolCommands?: string[];
-  inputSchema?: unknown;
-  outputDataSchema: z.ZodTypeAny;
-  annotations?: ToolAnnotations;
-  callback: (cbArgs: any) => unknown;
-}): void {
-  registerManifestEntry({
-    name: args.name,
-    description: args.description,
-    category: args.category,
-    protocolCommands: args.protocolCommands,
-    inputSchema: args.inputSchema,
-    outputDataSchema: args.outputDataSchema,
-    annotations: args.annotations,
-  });
-
-  // Wrap the callback to automatically measure wall-clock time.
-  // If the callback already sets elapsedMs on the envelope, that value
-  // is preserved. Otherwise elapsedMs is filled in automatically.
-  const timedCallback = async (toolArgs: any) => {
-    const startMs = performance.now();
-    const result = await args.callback(toolArgs);
-    const elapsed = Math.round(performance.now() - startMs);
-    if (
-      result &&
-      typeof result === "object" &&
-      "structuredContent" in result
-    ) {
-      const structuredContent = (result as any).structuredContent;
-      if (structuredContent && typeof structuredContent === "object" && structuredContent.elapsedMs === undefined) {
-        structuredContent.elapsedMs = elapsed;
-      }
-    }
-    return result;
-  };
-
-  args.server.registerTool(
-    args.name,
-    {
-      description: args.description,
-      inputSchema: args.inputSchema as never,
-      outputSchema: toolEnvelopeSchema(args.outputDataSchema),
-      annotations: args.annotations,
-    },
-    timedCallback as never,
-  );
-}
-
-export function registerTextTool(args: {
-  server: McpServer;
-  name: string;
-  description: string;
-  category: ToolCategory;
-  protocolCommands?: string[];
-  inputSchema?: unknown;
-  annotations?: ToolAnnotations;
-  outputDataSchema?: z.ZodTypeAny;
-  /**
-   * Optional session reference. When provided, the wrapper short-
-   * circuits the tool with an `unavailable` envelope if the session's
-   * most recent load classification is `type-error` — see §1.3 in the
-   * agent UX observations doc. Callers that are informational in
-   * nature (status tools, load-family, tools that MUST run in error
-   * state to surface the errors themselves) should omit the session
-   * argument so the wrapper doesn't gate them. If omitted, the
-   * behavior is unchanged from the pre-#39 release.
-   */
-  session?: AgdaSession;
-  callback: (cbArgs: any) => Promise<string>;
-}): void {
-  const outputDataSchema =
-    args.outputDataSchema ?? z.object({ text: z.string() });
-
-  registerStructuredTool({
-    server: args.server,
-    name: args.name,
-    description: args.description,
-    category: args.category,
-    protocolCommands: args.protocolCommands,
-    inputSchema: args.inputSchema,
-    annotations: args.annotations,
-    outputDataSchema,
-    callback: async (cbArgs: any) => {
-      const startMs = performance.now();
-      if (args.session) {
-        const unavailable = sessionErrorStateGate(args.session, args.name, { text: "" });
-        if (unavailable) return unavailable;
-      }
-      try {
-        const textValue = await args.callback(cbArgs);
-        return makeToolResult(
-          okEnvelope({
-            tool: args.name,
-            summary: textValue,
-            data: { text: textValue },
-            elapsedMs: Math.round(performance.now() - startMs),
-          }),
-          textValue,
-        );
-      } catch (err) {
-        return makeTextToolErrorResult(args.name, err, { text: "" });
-      }
-    },
-  });
-}
-
-export function registerGoalTextTool<A extends Record<string, unknown>>(args: {
-  server: McpServer;
-  session: AgdaSession;
-  name: string;
-  description: string;
-  category: ToolCategory;
-  protocolCommands?: string[];
-  inputSchema: unknown;
-  annotations?: ToolAnnotations;
-  outputDataSchema?: z.ZodTypeAny;
-  callback: (cbArgs: A & { goalId: number }) => Promise<string>;
-}): void {
-  const outputDataSchema =
-    args.outputDataSchema
-    ?? z.object({ text: z.string(), goalId: z.number() });
-
-  registerStructuredTool({
-    server: args.server,
-    name: args.name,
-    description: args.description,
-    category: args.category,
-    protocolCommands: args.protocolCommands,
-    inputSchema: args.inputSchema,
-    annotations: args.annotations,
-    outputDataSchema,
-    callback: async (cbArgs: A & { goalId: number }) => {
-      const startMs = performance.now();
-      const invalid = validateGoalId(args.session, cbArgs.goalId, args.name);
-      if (invalid) {
-        return invalid;
-      }
-
-      // §1.3: gate goal-based tools when the session's last load
-      // failed with type-error. Even with a valid goalId that was
-      // captured during the previous successful load, the session
-      // can no longer answer goal queries coherently and Agda will
-      // echo the stale error — surface it as `unavailable` instead.
-      const unavailable = sessionErrorStateGate(
-        args.session,
-        args.name,
-        { text: "", goalId: cbArgs.goalId },
-      );
-      if (unavailable) return unavailable;
-
-      try {
-        const warn = stalenessWarning(args.session);
-        const body = await args.callback(cbArgs);
-        const textValue = warn + body;
-        return makeToolResult(
-          okEnvelope({
-            tool: args.name,
-            summary: body,
-            data: { text: body, goalId: cbArgs.goalId },
-            stale: args.session.isFileStale() || undefined,
-            elapsedMs: Math.round(performance.now() - startMs),
-          }),
-          textValue,
-        );
-      } catch (err) {
-        return makeTextToolErrorResult(args.name, err, {
-          text: "",
-          goalId: cbArgs.goalId,
-        });
-      }
-    },
-  });
-}
+export {
+  registerGoalTextTool,
+  registerStructuredTool,
+  registerTextTool,
+  wrapGoalHandler,
+  wrapHandler,
+  wrapStructuredGoalHandler,
+  wrapStructuredHandler,
+} from "./tool-registration.js";
